requirements.txt
crewai[tools]
Pillow
pyyaml
google-generativeai
langchain
openai
ollama
streamlit